add sound clues to the game

+++ src/main/kotlin/com/arkhamusserver/arkhamus/model/database/entity/game/leveldesign/clues/SoundClue.kt
@@ -13,7 +13,7 @@ class SoundClue(
     var x: Double,
     var y: Double,
     var z: Double,
-    var inGameId: Int,
+    var inGameId: Long,
     @ManyToOne
     @JoinColumn(name = "levelId", nullable = false)
     var level: Level? = null,
