+++ src/main/kotlin/com/arkhamusserver/arkhamus/view/levelDesign/JsonSoundClueJammer.kt
@@ -1,10 +1,10 @@
 package com.arkhamusserver.arkhamus.view.levelDesign
 
 data class JsonSoundClueJammer(
-    var id: Int?,
+    var id: Long?,
     var interactionRadius: Double?,
     var name: String?,
-    var relatedClue: Int?,
+    var relatedClue: Long?,
     var x: Double?,
     var y: Double?,
     var z: Double?,
